Translated multiple files for Python3. Joining the "wmlgrammar.py" and "wmldata.py" files because there was a circular dependency.
