+++ data/tools/wesnoth_addon_manager
@@ -12,7 +12,6 @@ add-ons.
 import sys, os.path, re, time, glob, shutil
 from subprocess import Popen
 import wesnoth.wmldata as wmldata
-import wesnoth.wmlparser as wmlparser
 from wesnoth.campaignserver_client import CampaignClient
 
 if __name__ == "__main__":
@@ -32,7 +31,7 @@ if __name__ == "__main__":
         help="Output a HTML overview into the given directory.",)
     argumentparser.add_argument("-p", "--port",
         help="specify server port or BfW version (%s)" % " or ".join(
-        map(lambda x: x[1], CampaignClient.portmap)),
+        [x[1] for x in CampaignClient.portmap]),
         default=CampaignClient.portmap[0][0])
     argumentparser.add_argument("-l", "--list", help="list available add-ons",
         action="store_true",)
@@ -112,11 +111,11 @@ if __name__ == "__main__":
         while not mythread.event.isSet():
             mythread.event.wait(1)
             if pcounter != cs.counter:
-                print "%s: %d/%d" % (name, cs.counter, cs.length)
+                print("%s: %d/%d" % (name, cs.counter, cs.length))
                 pcounter = cs.counter
 
         if args.raw_download:
-            file(name, "w").write(mythread.data)
+            open(name, "w").write(mythread.data)
         else:
             decoded = cs.decode(mythread.data)
             dirname = os.path.join(cdir, name)
@@ -127,13 +126,13 @@ if __name__ == "__main__":
             try: os.remove(oldcfg_path)
             except OSError: pass
 
-            print "Unpacking %s..." % name
+            print("Unpacking %s..." % name)
             cs.unpackdir(decoded, cdir, verbose=args.verbose)
 
 
             info = os.path.join(dirname, "_info.cfg")
             try:
-                f = file(info, "w")
+                f = open(info, "w")
                 infowml = """#
 # File automatically generated by Wesnoth to keep track
 # of version information on installed add-ons. DO NOT EDIT!
@@ -152,7 +151,7 @@ if __name__ == "__main__":
             except OSError:
                 pass
             for message in decoded.find_all("message", "error"):
-                print message.get_text_val("message")
+                print(message.get_text_val("message"))
 
             if args.tar:
                 try: os.mkdir(args.tar)
@@ -179,7 +178,7 @@ if __name__ == "__main__":
         if not os.path.exists(name):
             return None, None
 
-        p = wmlparser.Parser(None)
+        p = wmldata.Parser(None)
         p.parse_file(name)
         info = wmldata.DataSub("WML")
         p.parse_top(info)
@@ -224,7 +223,7 @@ if __name__ == "__main__":
                         sys.stdout.write(f.ljust(column_sizes[i]))
                     sys.stdout.write("\n")
             for message in data.find_all("message", "error"):
-                print message.get_text_val("message")
+                print(message.get_text_val("message"))
         else:
             sys.stderr.write("Could not connect.\n")
 
@@ -256,31 +255,31 @@ if __name__ == "__main__":
                 if version != local_version or uploads > local_uploads:
                     get(name, title, version, type, uploads, dependencies, args.campaigns_dir)
                 else:
-                    print "Not downloading", name, \
+                    print("Not downloading", name, \
                         "as the version already is", local_version, \
-                        "(The add-on got re-uploaded.)"
+                        "(The add-on got re-uploaded.)")
             else:
                 if args.verbose:
-                    print "Not downloading", name, \
-                        "because it is already up-to-date."
+                    print("Not downloading", name, \
+                        "because it is already up-to-date.")
 
     elif args.unpack:
         cs = CampaignClient(address)
-        data = file(args.unpack).read()
+        data = open(args.unpack).read()
         decoded = cs.decode(data)
-        print "Unpacking %s..." % args.unpack
+        print("Unpacking %s..." % args.unpack)
         cs.unpackdir(decoded, args.campaigns_dir,  verbose=True)
     elif args.remove:
         cs = CampaignClient(address)
         data = cs.delete_campaign(args.remove, args.password)
         for message in data.find_all("message", "error"):
-            print message.get_text_val("message")
+            print(message.get_text_val("message"))
 
     elif args.change_passphrase:
         cs = CampaignClient(address)
         data = cs.change_passphrase(*args.change_passphrase)
         for message in data.find_all("message", "error"):
-            print message.get_text_val("message")
+            print(message.get_text_val("message"))
 
     elif args.upload:
         cs = CampaignClient(address)
@@ -341,11 +340,11 @@ if __name__ == "__main__":
         while not mythread.event.isSet():
             mythread.event.wait(1)
             if cs.counter != pcounter:
-                print "%d/%d" % (cs.counter, cs.length)
+                print("%d/%d" % (cs.counter, cs.length))
                 pcounter = cs.counter
 
         for message in mythread.data.find_all("message", "error"):
-            print message.get_text_val("message")
+            print(message.get_text_val("message"))
 
     elif args.update or args.status:
         if args.status:
