+++ data/tools/wmlvalidator
@@ -8,12 +8,12 @@ Use --help to see usage.
 #-define verbosity levels better
 
 import wesnoth.wmldata as wmldata
-import wesnoth.wmlparser as wmlparser
+
 import wesnoth.wmlgrammar as wmlgrammar
 import re
 
 def print_indent(string, depth, char=' '):
-    print "%s%s" % (depth * char, string)
+    print("%s%s" % (depth * char, string))
 
 class Validator:
     """
@@ -35,13 +35,13 @@ class Validator:
         else:
             verbosename = "%s (%s)" % (node.name, name)
 
-        if self.verbosity > 1:
-            print_indent(node.name, depth)
+        #if self.verbosity > 1:
+        #    print_indent(node.name, depth)
 
         try:
             schema = self.schema.get_element(name)
         except KeyError:
-            print "No valid schema found for %s" % verbosename
+            print("No valid schema found for %s" % verbosename)
             return
 
 
@@ -53,56 +53,51 @@ class Validator:
             matches = node.get_texts(attribute.name)
             nummatches = len(matches)
             if attribute.freq == wmlgrammar.REQUIRED and nummatches != 1:
-                print "(%s:%d) Attribute '[%s] %s' should appear exactly once, not %d times" % (node.file, node.line, verbosename, attribute.name, nummatches)
+                print("(%s:%d) Attribute '[%s] %s' should appear exactly once, not %d times" % (node.file, node.line, verbosename, attribute.name, nummatches))
             elif attribute.freq == wmlgrammar.OPTIONAL and nummatches > 1:
-                print "(%s:%d) Attribute '[%s] %s' should appear at most once, not %d times" % (node.file, node.line, verbosename, attribute.name, nummatches)
+                print("(%s:%d) Attribute '[%s] %s' should appear at most once, not %d times" % (node.file, node.line, verbosename, attribute.name, nummatches))
             elif attribute.freq == wmlgrammar.FORBIDDEN and nummatches > 0:
-                print "(%s:%d) Attribute '[%s] %s' should not appear. It appears %d times" % (node.file, node.line, verbosename, attribute.name, nummatches)
+                print("(%s:%d) Attribute '[%s] %s' should not appear. It appears %d times" % (node.file, node.line, verbosename, attribute.name, nummatches))
             for match in matches:
                 if 'translatable' in attribute.optionals and match.is_translatable() == False:
-                    print "(%s:%d) Attribute '[%s] %s's value is translatable, but haven't _ at the beginning" % (node.file, node.line, verbosename, attribute.name)
+                    print("(%s:%d) Attribute '[%s] %s's value is translatable, but haven't _ at the beginning" % (node.file, node.line, verbosename, attribute.name))
                 elif 'translatable' not in attribute.optionals and match.is_translatable() == True:
-                    print "(%s:%d) Attribute '[%s] %s's value isn't translatable, but have a _ at the beginning" % (node.file, node.line, verbosename, attribute.name)
+                    print("(%s:%d) Attribute '[%s] %s's value isn't translatable, but have a _ at the beginning" % (node.file, node.line, verbosename, attribute.name))
                 if 'list' in attribute.optionals:
                     pos = 1
                     for i in match.data.split(","):
                         if i[0] == ' ': i = i[1:]
                         if not attribute.validate(i):
-                            print "(%s:%d) Attribute '[%s] %s's value in list should be %s, found at position %d: %s" % (node.file, node.line, verbosename, attribute.name, attribute.type, pos, i)
+                            print("(%s:%d) Attribute '[%s] %s's value in list should be %s, found at position %d: %s" % (node.file, node.line, verbosename, attribute.name, attribute.type, pos, i))
                         pos += 1
                 else:
                     if not attribute.validate(match.data):
-                        print "(%s:%d) Attribute '[%s] %s's value should be %s, found: %s" % (node.file, node.line, verbosename, attribute.name, attribute.type, match.data)
+                        print("(%s:%d) Attribute '[%s] %s's value should be %s, found: %s" % (node.file, node.line, verbosename, attribute.name, attribute.type, match.data))
                 node.remove(match) # Get rid of these so we can see what's left
         for attribute in node.get_all_text():
-            print "(%s:%d) Attribute '[%s] %s' found, which has no meaning there" % (node.file, node.line, verbosename, attribute.name)
+            print("(%s:%d) Attribute '[%s] %s' found, which has no meaning there" % (node.file, node.line, verbosename, attribute.name))
 
         # Validate the elements
         for element in schema.get_elements():
             matches = node.get_subs(element.name)
             nummatches = len(matches)
             if element.freq == wmlgrammar.REQUIRED and nummatches != 1:
-                print "(%s:%d) Element '[%s] [%s]' should appear exactly once, not %d times" % (node.file, node.line, verbosename, element.name, nummatches)
+                print("(%s:%d) Element '[%s] [%s]' should appear exactly once, not %d times" % (node.file, node.line, verbosename, element.name, nummatches))
             elif element.freq == wmlgrammar.OPTIONAL and nummatches > 1:
-                print "(%s:%d) Element '[%s] [%s]' should appear at most once, not %d times" % (node.file, node.line, verbosename, element.name, nummatches)
+                print("(%s:%d) Element '[%s] [%s]' should appear at most once, not %d times" % (node.file, node.line, verbosename, element.name, nummatches))
             for match in matches:
                 self.validate(match, depth+1, element.subname)
                 node.remove(match)
 
         for element in node.get_all_subs():
-            print "(%s:%d) Element '[%s] [%s]' found, which has no meaning there" % (node.file, node.line, verbosename, element.name)
+            print("(%s:%d) Element '[%s] [%s]' found, which has no meaning there" % (node.file, node.line, verbosename, element.name))
             # Do we want to do this?
             if False:
-                print "Attempting to validate [%s] anyway" % element.name
+                print("Attempting to validate [%s] anyway" % element.name)
                 self.validate(element, depth+1)
 
 if __name__ == '__main__':
     import argparse, subprocess, os, codecs, sys
-
-    # Ugly hack to force the output of UTF-8.
-    # This prevents us from crashing when we're being verbose
-    #  and encounter a non-ascii character.
-    sys.stdout = codecs.getwriter('utf-8')(sys.stdout)
 
     ap = argparse.ArgumentParser("Usage: %(prog)s [options]")
     ap.add_argument("-p", "--path",
@@ -141,17 +136,17 @@ if __name__ == '__main__':
     if len(args.filename) < 1:
         args.filename.append(os.path.join(args.path, '_main.cfg'))
 
-    if args.verbose > 1:
-        print "Args: %s\n"% (args, )
+    #if args.verbose > 1:
+    #    print("Args: %s\n"% (args, ))
 
     if not args.schema:
         args.schema = os.path.join(args.path, 'schema.cfg')
 
     # Parse the schema
-    parser = wmlparser.Parser(args.path)
+    parser = wmldata.Parser(args.path)
 
-    if args.verbose > 3:
-        parser.verbose = True
+    #if args.verbose > 3:
+    #    parser.verbose = True
     parser.parse_file(args.schema)
 
     schema = wmldata.DataSub("schema")
@@ -161,10 +156,10 @@ if __name__ == '__main__':
     validator = Validator(schema, args.verbose)
 
     # Parse the WML
-    parser = wmlparser.Parser(args.path, args.userpath)
+    parser = wmldata.Parser(args.path, args.userpath)
 
-    if args.verbose > 3:
-        parser.verbose = True
+    #if args.verbose > 3:
+    #    parser.verbose = True
 
     if args.userpath:
         parser.parse_text("{~add-ons}")
